Remove "Talk to Us" button for simplicity
This commit removes the "Talk to Us" button, which links to the
repository's GitHub issues. The Slack page is probably a better place to
get in touch and talk with us.

This commit also makes it so that on a desktop, the two remaining
buttons to be side to side, rather than under each other. However, on
mobile, the buttons will be on top of each other.